Add 'semester' table to db schema
Update 'course_schedules' seeding to use new semester table and seed for more semesters. Ensure that in each semester, courses from the same department and with the same 'YearOfCourse' value do not have date&time conflicts.

Update 'enrollment' seeding to use new semester data. Ensure that in each semester, a student enrolls in the courses compatible with the grade the student is during that semester.

Update 'exams' seding to use new semester data. Ensure that the exam dates in each semester is compatible with the start date of that semester.